setup-nushell.sh OSTYPE check

+++ .config/nushell/setup-nushell.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-rm -rf "$HOME/Library/Application Support/nushell" && ln -sf "$HOME/.dotfiles/.config/nushell" "$HOME/Library/Application Support/nushell"
+if [[ "$OSTYPE" == "darwin"* ]]; then
+	rm -rf "$HOME/Library/Application Support/nushell" && ln -sf "$HOME/.dotfiles/.config/nushell" "$HOME/Library/Application Support/nushell"
+fi
 
 NU_PATH="$(which nu)"
 echo "$NU_PATH" | sudo tee -a /etc/shells
